Tweaked numbers and added red boss(no AI)

+++ Assets/Scripts/Enemy/Enemy.cs
@@ -9,7 +9,7 @@ public abstract class Enemy : MonoBehaviour {
     [Header("• Properties")]
     public float Speed;
     public float Health;
-    public float HitPower;
+    public int HitPower;
     public bool CanJump;
 
     [Space]
